Change the way the topics are stored and saved
The saving of the topic modeling results was relying on the model to
enter the data in the correct way in order to properly work. This is
impractical, especially when trying out various topic models.
Therefore, the new classes are designed in such a way that data can
be easily added by the model. They are themselved not part of the
saving procedure, but provide functions to export the data to the
saving module.