Update tokenizer and embedding model from TOMEF
The tokenizer and the word2vec embedding class have been updated to
reflect the last changes made to the respective modules in TOMEF.
This includes a faster filtering of words and removing the dependency
on the default tokenizer.

Additionally, the embedding data is now loaded directly from s3 bucket
if it is not available locally.